fix warnings about non-matching sign in operations

+++ openvpn/compress/lz4.hpp
@@ -71,11 +71,11 @@ namespace openvpn {
 	}
 
       // do compress
-      const int comp_size = LZ4_compress_default((const char *)buf.c_data(), (char *)work.data(),
-                                                 (int)buf.size(), (int)work.capacity());
+      const ssize_t comp_size = LZ4_compress_default((const char *)buf.c_data(), (char *)work.data(),
+						     (int)buf.size(), (int)work.capacity());
 
       // did compression actually reduce data length?
-      if (comp_size < buf.size())
+      if (comp_size < (ssize_t)buf.size())
 	{
 	  if (comp_size <= 0)
 	    {
